pyproject.toml
[tool.poetry]
name = "taskik"
version = "0.1.0"
description = ""
authors = ["N1ntek <zobalarm@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pydantic = {extras = ["email"], version = "^2.7.1"}
bcrypt = "^4.1.3"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
include = '\.pyi?$'